dumper: do never repeat request information, traceback or SQL query if unchanged
The body of a log consists in 3 pieces of information:
- request
- traceback
- SQL Query (sometimes)

This commit goes further than 22d7f0905a18f1445fa785ac4e0cba486e7550c2,
by not repeating any of the above 3 parts if any other changes.

This is particularly useful for the request information or the SQL query
because they can be huge while remaining the same for a long time.